Use AdtDefinedHere subdiagnostic
