+++ tests/ui/pattern/suggest-adding-appropriate-missing-pattern-excluding-comments.stderr
@@ -4,11 +4,6 @@ error[E0004]: match is non-exhaustive
 LL |     match Some(1) {
    |           ^^^^^^^ pattern `None` not covered
    |
-note: `Option<i32>` defined here
-  --> $SRC_DIR/core/src/option.rs:LL:COL
-  ::: $SRC_DIR/core/src/option.rs:LL:COL
-   |
-   = note: not covered
    = note: the matched value is of type `Option<i32>`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
