+++ tests/ui/pattern/usefulness/doc-hidden-non-exhaustive.stderr
@@ -4,11 +4,6 @@ error[E0004]: match is non-exhaustive
 LL |     match HiddenEnum::A {
    |           ^^^^^^^^^^^^^ pattern `_` not covered
    |
-note: `HiddenEnum` defined here
-  --> $DIR/auxiliary/hidden.rs:1:1
-   |
-LL | pub enum HiddenEnum {
-   | ^^^^^^^^^^^^^^^^^^^
    = note: the matched value is of type `HiddenEnum`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
@@ -22,14 +17,6 @@ error[E0004]: match is non-exhaustive
 LL |     match HiddenEnum::A {
    |           ^^^^^^^^^^^^^ pattern `HiddenEnum::B` not covered
    |
-note: `HiddenEnum` defined here
-  --> $DIR/auxiliary/hidden.rs:3:5
-   |
-LL | pub enum HiddenEnum {
-   | -------------------
-LL |     A,
-LL |     B,
-   |     ^ not covered
    = note: the matched value is of type `HiddenEnum`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
@@ -43,14 +30,6 @@ error[E0004]: match is non-exhaustive
 LL |     match HiddenEnum::A {
    |           ^^^^^^^^^^^^^ patterns `HiddenEnum::B` and `_` not covered
    |
-note: `HiddenEnum` defined here
-  --> $DIR/auxiliary/hidden.rs:3:5
-   |
-LL | pub enum HiddenEnum {
-   | -------------------
-LL |     A,
-LL |     B,
-   |     ^ not covered
    = note: the matched value is of type `HiddenEnum`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern, a match arm with multiple or-patterns as shown, or multiple match arms
    |
@@ -64,11 +43,6 @@ error[E0004]: match is non-exhaustive
 LL |     match None {
    |           ^^^^ patterns `Some(HiddenEnum::B)` and `Some(_)` not covered
    |
-note: `Option<HiddenEnum>` defined here
-  --> $SRC_DIR/core/src/option.rs:LL:COL
-  ::: $SRC_DIR/core/src/option.rs:LL:COL
-   |
-   = note: not covered
    = note: the matched value is of type `Option<HiddenEnum>`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern, a match arm with multiple or-patterns as shown, or multiple match arms
    |
@@ -83,13 +57,13 @@ LL |     match InCrate::A {
    |           ^^^^^^^^^^ pattern `InCrate::C` not covered
    |
 note: `InCrate` defined here
-  --> $DIR/doc-hidden-non-exhaustive.rs:11:5
+  --> $DIR/doc-hidden-non-exhaustive.rs:7:6
    |
 LL | enum InCrate {
-   |      -------
+   |      ^^^^^^^
 ...
 LL |     C,
-   |     ^ not covered
+   |     - not covered
    = note: the matched value is of type `InCrate`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
