+++ tests/ui/pattern/usefulness/issue-35609.stderr
@@ -105,8 +105,6 @@ error[E0004]: match is non-exhaustive
 LL |     match Some(A) {
    |           ^^^^^^^ patterns `Some(Enum::B)`, `Some(Enum::C)`, `Some(Enum::D)` and 2 more not covered
    |
-note: `Option<Enum>` defined here
-  --> $SRC_DIR/core/src/option.rs:LL:COL
    = note: the matched value is of type `Option<Enum>`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern as shown, or multiple match arms
    |
