+++ tests/ui/pattern/usefulness/issue-3601.stderr
@@ -4,8 +4,6 @@ error[E0004]: match is non-exhaustive
 LL |         box NodeKind::Element(ed) => match ed.kind {
    |                                            ^^^^^^^ pattern `box _` not covered
    |
-note: `Box<ElementKind>` defined here
-  --> $SRC_DIR/alloc/src/boxed.rs:LL:COL
    = note: the matched value is of type `Box<ElementKind>`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
