+++ tests/ui/pattern/usefulness/stable-gated-patterns.stderr
@@ -4,14 +4,6 @@ error[E0004]: match is non-exhaustive
 LL |     match UnstableEnum::Stable {
    |           ^^^^^^^^^^^^^^^^^^^^ patterns `UnstableEnum::Stable2` and `_` not covered
    |
-note: `UnstableEnum` defined here
-  --> $DIR/auxiliary/unstable.rs:9:5
-   |
-LL | pub enum UnstableEnum {
-   | ---------------------
-...
-LL |     Stable2,
-   |     ^^^^^^^ not covered
    = note: the matched value is of type `UnstableEnum`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern, a match arm with multiple or-patterns as shown, or multiple match arms
    |
@@ -25,11 +17,6 @@ error[E0004]: match is non-exhaustive
 LL |     match UnstableEnum::Stable {
    |           ^^^^^^^^^^^^^^^^^^^^ pattern `_` not covered
    |
-note: `UnstableEnum` defined here
-  --> $DIR/auxiliary/unstable.rs:5:1
-   |
-LL | pub enum UnstableEnum {
-   | ^^^^^^^^^^^^^^^^^^^^^
    = note: the matched value is of type `UnstableEnum`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
