+++ tests/ui/pattern/usefulness/unstable-gated-patterns.stderr
@@ -4,14 +4,6 @@ error[E0004]: match is non-exhaustive
 LL |     match UnstableEnum::Stable {
    |           ^^^^^^^^^^^^^^^^^^^^ pattern `UnstableEnum::Unstable` not covered
    |
-note: `UnstableEnum` defined here
-  --> $DIR/auxiliary/unstable.rs:11:5
-   |
-LL | pub enum UnstableEnum {
-   | ---------------------
-...
-LL |     Unstable,
-   |     ^^^^^^^^ not covered
    = note: the matched value is of type `UnstableEnum`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
