+++ tests/ui/rfc-2008-non-exhaustive/enum.stderr
@@ -23,11 +23,6 @@ error[E0004]: match is non-exhaustive
 LL |     match enum_unit {
    |           ^^^^^^^^^ pattern `_` not covered
    |
-note: `NonExhaustiveEnum` defined here
-  --> $DIR/auxiliary/enums.rs:4:1
-   |
-LL | pub enum NonExhaustiveEnum {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: the matched value is of type `NonExhaustiveEnum`, which is marked as non-exhaustive
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
@@ -41,11 +36,6 @@ error[E0004]: match is non-exhaustive
 LL |     match enum_unit {};
    |           ^^^^^^^^^ pattern `_` not covered
    |
-note: `NonExhaustiveEnum` defined here
-  --> $DIR/auxiliary/enums.rs:4:1
-   |
-LL | pub enum NonExhaustiveEnum {
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: the matched value is of type `NonExhaustiveEnum`, which is marked as non-exhaustive
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
