+++ tests/ui/rfc-2008-non-exhaustive/uninhabited/match_with_exhaustive_patterns.stderr
@@ -61,15 +61,6 @@ error[E0004]: match is non-exhaustive
 LL |     match x {}
    |           ^ patterns `UninhabitedVariants::Tuple(_)` and `UninhabitedVariants::Struct { .. }` not covered
    |
-note: `UninhabitedVariants` defined here
-  --> $DIR/auxiliary/uninhabited.rs:17:23
-   |
-LL | pub enum UninhabitedVariants {
-   | ----------------------------
-LL |     #[non_exhaustive] Tuple(!),
-   |                       ^^^^^ not covered
-LL |     #[non_exhaustive] Struct { x: ! }
-   |                       ^^^^^^ not covered
    = note: the matched value is of type `UninhabitedVariants`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern, a match arm with multiple or-patterns as shown, or multiple match arms
    |
