+++ tests/ui/uninhabited/uninhabited-matches-feature-gated.stderr
@@ -4,11 +4,6 @@ error[E0004]: match is non-exhaustive
 LL |     let _ = match x {
    |                   ^ pattern `Err(_)` not covered
    |
-note: `Result<u32, &Void>` defined here
-  --> $SRC_DIR/core/src/result.rs:LL:COL
-  ::: $SRC_DIR/core/src/result.rs:LL:COL
-   |
-   = note: not covered
    = note: the matched value is of type `Result<u32, &Void>`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
@@ -83,11 +78,6 @@ error[E0004]: match is non-exhaustive
 LL |     let _ = match x {
    |                   ^ pattern `Err(_)` not covered
    |
-note: `Result<u32, Void>` defined here
-  --> $SRC_DIR/core/src/result.rs:LL:COL
-  ::: $SRC_DIR/core/src/result.rs:LL:COL
-   |
-   = note: not covered
    = note: the matched value is of type `Result<u32, Void>`
 help: ensure that all possible cases are being handled by adding a match arm with a wildcard pattern or an explicit pattern as shown
    |
